fix: rename repo and rescope stacks
