+++ carpark/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@web3-storage/upload-api-carpark",
+  "name": "@web3-storage/w3infra-carpark",
   "version": "0.0.0",
   "type": "module",
   "scripts": {
