+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "upload-api",
+  "name": "w3infra",
   "version": "0.0.0",
   "private": true,
   "scripts": {
@@ -10,7 +10,7 @@
     "console": "sst console",
     "lint": "tsc && eslint '**/*.js'",
     "clean": "rm -rf docs node_modules pnpm-lock.yaml packages/*/{pnpm-lock.yaml,coverage,.nyc_output,dist,node_modules,.next}",
-    "test": "npm test -w api -w carpark -w replicator -w satnav"
+    "test": "npm test -w upload-api -w carpark -w replicator -w satnav"
   },
   "devDependencies": {
     "@serverless-stack/cli": "^1.18.0",
@@ -60,9 +60,9 @@
     }
   },
   "workspaces": [
-    "api",
     "carpark",
     "replicator",
-    "satnav"
+    "satnav",
+    "upload-api"
   ]
 }
