+++ replicator/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@web3-storage/upload-api-replicator",
+  "name": "@web3-storage/w3infra-replicator",
   "version": "0.0.0",
   "type": "module",
   "scripts": {
