+++ satnav/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@web3-storage/upload-api-satnav",
+  "name": "@web3-storage/w3infra-satnav",
   "version": "0.0.0",
   "type": "module",
   "scripts": {
