+++ sst.json
@@ -1,5 +1,5 @@
 {
-  "name": "upload-api",
+  "name": "w3infra",
   "region": "us-west-2",
   "main": "stacks/index.js"
 }
