+++ stacks/index.js
@@ -1,6 +1,7 @@
 import { Tags } from 'aws-cdk-lib'
 
-import { ApiStack } from './api-stack.js'
+import { UploadApiStack } from './upload-api-stack.js'
+import { UploadDbStack } from './upload-db-stack.js'
 import { BusStack } from './bus-stack.js'
 import { CarparkStack } from './carpark-stack.js'
 import { SatnavStack } from './satnav-stack.js'
@@ -17,13 +18,14 @@ export default function (app) {
   })
   app.stack(BusStack)
   app.stack(CarparkStack)
-  app.stack(ApiStack)
+  app.stack(UploadDbStack)
   app.stack(SatnavStack)
+  app.stack(UploadApiStack)
 
   // tags let us discover all the aws resource costs incurred by this app
   // see: https://docs.sst.dev/advanced/tagging-resources
-  Tags.of(app).add('Project', 'upload-api')
-  Tags.of(app).add('Repository', 'https://github.com/web3-storage/upload-api')
+  Tags.of(app).add('Project', 'w3infra')
+  Tags.of(app).add('Repository', 'https://github.com/web3-storage/w3infra')
   Tags.of(app).add('Environment', `${app.stage}`)
   Tags.of(app).add('ManagedBy', 'SST')
 }
