+++ upload-api/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@web3-storage/upload-api",
+  "name": "@web3-storage/w3infra-upload-api",
   "version": "3.0.0",
   "type": "module",
   "scripts": {
