+++ upload-api/service/store/remove.js
