+++ upload-api/service/upload/remove.js
